places req.user

+++ src/modules/places/places.service.ts
@@ -1,4 +1,9 @@
-import { BadRequestException, Injectable, Logger } from '@nestjs/common'
+import {
+  BadRequestException,
+  Injectable,
+  Logger,
+  UnauthorizedException
+} from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Place } from '../../entities/place.entity'
 import { Repository } from 'typeorm'
@@ -25,16 +30,17 @@ export class PlacesService {
       })
     } catch (err) {
       console.log(err.message)
-      throw new BadRequestException(
-        'Error while searching for all locations.'
-      )
+      throw new BadRequestException('Error while searching for all locations.')
     } finally {
       this.logger.log('Searching for all locations.')
     }
   }
 
-  async getRecent(id: number): Promise<Place[]> {
+  async getRecent(id: number, user_id: number): Promise<Place[]> {
     try {
+      if (id !== user_id) {
+        throw Error
+      }
       const result: Place[] = []
       const places: Place[] = await this.placesRepository.find({
         order: { updated_at: 'DESC' },
@@ -56,8 +62,14 @@ export class PlacesService {
     }
   }
 
-  async getPersonalBest(user_id: number): Promise<IPersonalBest[]> {
+  async getPersonalBest(
+    user_id: number,
+    req: number
+  ): Promise<IPersonalBest[]> {
     try {
+      if (user_id !== req) {
+        throw Error
+      }
       const result: IPersonalBest[] = []
       const guesses = await this.guessRepository.find({ user_id: user_id })
       for (let i = 0; i < guesses.length; i++) {
@@ -99,14 +111,16 @@ export class PlacesService {
     }
   }
 
-  async createLocation(createLocationDto: CreateLocationDto): Promise<Place> {
+  async createLocation(
+    createLocationDto: CreateLocationDto,
+    user_id: number
+  ): Promise<Place> {
     try {
-      const user = await this.usersRepository.findOne(
-        createLocationDto.user_id,
-        { relations: ['places'] }
-      )
+      const user = await this.usersRepository.findOne(user_id, {
+        relations: ['places']
+      })
       const newLocation = this.placesRepository.create({
-        user_id: createLocationDto.user_id,
+        user_id: user_id,
         city: createLocationDto.address,
         lat: createLocationDto.lat,
         long: createLocationDto.long,
@@ -152,7 +166,10 @@ export class PlacesService {
     }
   }
 
-  async guessLocation(createGuessDto: CreateGuessDto): Promise<Guess> {
+  async guessLocation(
+    createGuessDto: CreateGuessDto,
+    user_id: number
+  ): Promise<Guess> {
     try {
       const location = await this.placesRepository.findOne(
         createGuessDto.location_id,
@@ -161,7 +178,7 @@ export class PlacesService {
       const guesses = await this.guessRepository.find()
       for (let i = 0; i < guesses.length; i++) {
         if (
-          createGuessDto.user_id === guesses[i].user_id &&
+          user_id === guesses[i].user_id &&
           createGuessDto.location_id === guesses[i].location_id
         ) {
           await this.guessRepository.remove(guesses[i])
@@ -169,7 +186,7 @@ export class PlacesService {
       }
       const newGuess = this.guessRepository.create({
         location_id: createGuessDto.location_id,
-        user_id: createGuessDto.user_id,
+        user_id: user_id,
         address: createGuessDto.address,
         distance: createGuessDto.distance
       })
@@ -185,20 +202,30 @@ export class PlacesService {
     }
   }
 
-  async deleteLocation(id: number): Promise<Place> {
+  async deleteLocation(id_param: number, user_id: number): Promise<Place> {
     try {
-      const location: Place = await this.placesRepository.findOne(id)
+      const l = await this.placesRepository.findOne(id_param)
+      if (l.user_id !== user_id) {
+        throw Error
+      }
+      const location: Place = await this.placesRepository.findOne(user_id)
       return this.placesRepository.remove(location)
     } catch (err) {
       console.log(err)
-      throw new BadRequestException(`Cannot delete a location with id: ${id}`)
+      throw new BadRequestException(
+        `Cannot delete a location with id: ${id_param}`
+      )
     } finally {
-      this.logger.log(`Deleting a location with id: ${id}`)
+      this.logger.log(`Deleting a location with id: ${id_param}`)
     }
   }
 
-  async deleteGuess(id: number): Promise<Guess> {
+  async deleteGuess(id: number, user_id: number): Promise<Guess> {
     try {
+      const g = await this.guessRepository.findOne(id)
+      if (g.user_id !== user_id) {
+        throw Error
+      }
       const guess: Guess = await this.guessRepository.findOne(id)
       return this.guessRepository.remove(guess)
     } catch (err) {
